+++ MomentumDiffusion.pxd
@@ -8,7 +8,7 @@ cimport ParallelMPI
 cdef class MomentumDiffusion:
     cdef:
         double [:] flux
-        int n_fluxes
+        Py_ssize_t n_fluxes
     cpdef initialize(self, Grid.Grid Gr, PrognosticVariables.PrognosticVariables PV,
                      DiagnosticVariables.DiagnosticVariables DV, ParallelMPI.ParallelMPI Pa)
     cpdef update(self, Grid.Grid Gr,ReferenceState.ReferenceState Rs, PrognosticVariables.PrognosticVariables PV,
