+++ ScalarAdvection.pxd
@@ -7,7 +7,7 @@ cdef class ScalarAdvection:
 
     cdef:
         double [:] flux
-        long order
+        Py_ssize_t order
 
     cpdef initialize(self,Grid.Grid Gr, PrognosticVariables.PrognosticVariables PV)
     cpdef update_cython(self, Grid.Grid Gr, ReferenceState.ReferenceState Rs,PrognosticVariables.PrognosticVariables PV, ParallelMPI.ParallelMPI Pa)
