+++ SparseSolvers.pxd
@@ -2,7 +2,7 @@ cdef class TDMA:
 
     cdef:
         double* scratch
-        long n
-        void initialize(self, long n)
+        Py_ssize_t n
+        void initialize(self, Py_ssize_t n)
         inline void solve(self, double* x, double* a, double* b, double* c) nogil
         void destroy(self)
